requirements.txt
# The Handy Controller Dependencies

# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0

# Development dependencies (optional)
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0
flake8>=3.9.0
isort>=5.9.0

# Documentation (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
